# Use google benchmarks and fix minor performance issues