--- a/.travis.yml
+++ b/.travis.yml
@@ -28,7 +28,7 @@ matrix:
           sources: [ 'ubuntu-toolchain-r-test', ]
           packages: [ 'libstdc++-5-dev', 'libxi-dev','libglu1-mesa-dev','x11proto-randr-dev','x11proto-xext-dev','libxrandr-dev','x11proto-xf86vidmode-dev','libxxf86vm-dev','libxcursor-dev','libxinerama-dev']
     - os: osx
-      osx_image: xcode7.3
+      osx_image: xcode12
       env: BUILDTYPE=Release
 
 cache: apt