--- a/bench/main.cpp
+++ b/bench/main.cpp
@@ -1,0 +1,7 @@
+#include <benchmark/benchmark.h>
+
+int main(int argc, char* argv[]) {
+    ::benchmark::Initialize(&argc, argv);
+    ::benchmark::RunSpecifiedBenchmarks();
+    return 0;
+}