--- a/include/mapbox/geojsonvt/tile.hpp
+++ b/include/mapbox/geojsonvt/tile.hpp
@@ -48,6 +48,7 @@ public:
           sq_tolerance(tolerance_ * tolerance_),
           lineMetrics(lineMetrics_) {
 
+        tile.features.reserve(source.size());
         for (const auto& feature : source) {
             const auto& geom = feature.geometry;
             const auto& props = feature.properties;
@@ -69,12 +70,12 @@ public:
 
 private:
     void addFeature(const vt_empty& empty, const property_map& props, const identifier& id) {
-        tile.features.push_back({ transform(empty), props, id });
+        tile.features.emplace_back(transform(empty), props, id);
     }
 
     void
     addFeature(const vt_point& point, const property_map& props, const identifier& id) {
-        tile.features.push_back({ transform(point), props, id });
+        tile.features.emplace_back(transform(point), props, id);
     }
 
     void addFeature(const vt_line_string& line,
@@ -84,11 +85,11 @@ private:
         if (!new_line.empty()) {
             if (lineMetrics) {
                 property_map newProps = props;
-                newProps["mapbox_clip_start"] = line.segStart / line.dist;
-                newProps["mapbox_clip_end"] = line.segEnd / line.dist;
-                tile.features.push_back({ std::move(new_line), newProps, id });
+                newProps.emplace(std::make_pair<std::string, value>("mapbox_clip_start", line.segStart / line.dist));
+                newProps.emplace(std::make_pair<std::string, value>("mapbox_clip_end", line.segEnd / line.dist));
+                tile.features.emplace_back(std::move(new_line), std::move(newProps), id);
             } else
-                tile.features.push_back({ std::move(new_line), props, id });
+                tile.features.emplace_back(std::move(new_line), props, id);
         }
     }
 
@@ -97,7 +98,7 @@ private:
                     const identifier& id) {
         const auto new_polygon = transform(polygon);
         if (!new_polygon.empty())
-            tile.features.push_back({ std::move(new_polygon), props, id });
+            tile.features.emplace_back(std::move(new_polygon), props, id);
     }
 
     void addFeature(const vt_geometry_collection& collection,
@@ -119,10 +120,10 @@ private:
         case 0:
             break;
         case 1:
-            tile.features.push_back({ std::move(new_multi[0]), props, id });
+            tile.features.emplace_back(std::move(new_multi[0]), props, id);
             break;
         default:
-            tile.features.push_back({ std::move(new_multi), props, id });
+            tile.features.emplace_back(std::move(new_multi), props, id);
             break;
         }
     }
@@ -141,7 +142,7 @@ private:
         mapbox::geometry::multi_point<int16_t> result;
         result.reserve(points.size());
         for (const auto& p : points) {
-            result.push_back(transform(p));
+            result.emplace_back(transform(p));
         }
         return result;
     }
@@ -149,9 +150,10 @@ private:
     mapbox::geometry::line_string<int16_t> transform(const vt_line_string& line) {
         mapbox::geometry::line_string<int16_t> result;
         if (line.dist > tolerance) {
+            result.reserve(line.size());
             for (const auto& p : line) {
                 if (p.z > sq_tolerance)
-                    result.push_back(transform(p));
+                    result.emplace_back(transform(p));
             }
         }
         return result;
@@ -160,9 +162,10 @@ private:
     mapbox::geometry::linear_ring<int16_t> transform(const vt_linear_ring& ring) {
         mapbox::geometry::linear_ring<int16_t> result;
         if (ring.area > sq_tolerance) {
+            result.reserve(ring.size());
             for (const auto& p : ring) {
                 if (p.z > sq_tolerance)
-                    result.push_back(transform(p));
+                    result.emplace_back(transform(p));
             }
         }
         return result;
@@ -170,28 +173,31 @@ private:
 
     mapbox::geometry::multi_line_string<int16_t> transform(const vt_multi_line_string& lines) {
         mapbox::geometry::multi_line_string<int16_t> result;
+        result.reserve(lines.size());
         for (const auto& line : lines) {
             if (line.dist > tolerance)
-                result.push_back(transform(line));
+                result.emplace_back(transform(line));
         }
         return result;
     }
 
     mapbox::geometry::polygon<int16_t> transform(const vt_polygon& rings) {
         mapbox::geometry::polygon<int16_t> result;
+        result.reserve(rings.size());
         for (const auto& ring : rings) {
             if (ring.area > sq_tolerance)
-                result.push_back(transform(ring));
+                result.emplace_back(transform(ring));
         }
         return result;
     }
 
     mapbox::geometry::multi_polygon<int16_t> transform(const vt_multi_polygon& polygons) {
         mapbox::geometry::multi_polygon<int16_t> result;
+        result.reserve(polygons.size());
         for (const auto& polygon : polygons) {
             const auto p = transform(polygon);
             if (!p.empty())
-                result.push_back(std::move(p));
+                result.emplace_back(std::move(p));
         }
         return result;
     }